Add custom CPU model when disabling KVM
When disabling KVM support and using the default CPU model, an error
will occur `1GiB pages not supported`. We can fix that by speficying the
cpu model in the run command.

We should go for the `max` cpu model, since it should support
everything, but note that this will fail on some qemu versions (e.g.
8.0.2), and it will have to be changed to something more specific, like
`Skylake-Server`

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>