run.sh: Use security model none for the root filesystem
The `passthrough` security model will attempt to create files on the
host using the same uid as the one on the guest, by calling `fchown`,
which will fail if `qemu` is not called as root.

Fix that by using security model `none`, which will still call the
`fchown`, but ignore the returned error.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>